backend/requirements.txt
# Core Dependencies
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.10
pydantic-settings==2.11.0
python-multipart==0.0.9
python-dotenv==1.1.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.21.0

# Authentication
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data Processing
python-dateutil==2.9.0
pytz==2025.1

# API Utils
httpx==0.28.1
python-json-logger==2.0.7
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
factory-boy==3.3.3
faker==33.0.0